refactor: Вынесена функция по созданию сессии в отдельный модуль
В боте использование функции убрано, до тех пор пока не будет доработано
использование базы
